add fonts and fix cycle dependency
