+++ App.js
@@ -1,10 +1,28 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import Navigator from './routes/routeStack';
+import { AppLoading } from 'expo';
+import * as Font from 'expo-font';
+
+const getFonts = () => {
+  return Font.loadAsync({
+    'encodeSansExpanded': require('./assets/fonts/EncodeSansExpanded-Regular.ttf')
+  })
+};
 
 //Navigator is an alias, we can give it any name we want, since it is a default export.
 export default function App() {
   console.disableYellowBox = true;
-  return (
-    <Navigator />
-  );
+  const [isReady, setIsReady] = React.useState(false);
+
+  if(!isReady) {
+    return(
+      <AppLoading
+        startAsync={getFonts}
+        onFinish={()=> setIsReady(true)}
+      />
+    )
+  } else {
+    return <Navigator />
+  }
+
 }
