+++ package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@react-native-community/masked-view": "^0.1.6",
     "expo": "~36.0.0",
+    "expo-font": "^8.0.0",
     "galio-framework": "^0.6.3",
     "geolib": "^3.2.1",
     "native-base": "^2.13.8",
