+++ react-native.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    assets: ['./assets/fonts/']
+};
